media-sound/pulseaudio-daemon: Fix failure removing qpaeq
Closes: https://bugs.gentoo.org/888131

+++ media-sound/pulseaudio-daemon/pulseaudio-daemon-16.1-r5.ebuild
@@ -253,7 +253,7 @@ src_install() {
 	meson_src_install
 
 	# Upstream installs qpaeq if fftw is found, we only want it with USE equalizer
-	if ! use equalizer; then
+	if ! use equalizer && [[ -f "${ED}"/usr/bin/qpaeq ]]; then
 		rm "${ED}"/usr/bin/qpaeq || die
 	fi
